Add UniformBinning to schemav2
When defining a Binning, the `edges` attribute can now
be either a list of floats (like before) or an instance
of UniformBinning(n, low, high). When using the latter
option, the Binning correction is optimized to use a
simple arithmetic formula rather than binary search to
look up the bin index.