Make interfaces of C++ correction types more uniform
Before this commit, most correction types implemented
an `evaluate` method while some implemented a `child`
method. When walking a correction's compute graph,
the `node_evaluate` visitor had to know which ones
required recursive visiting and which did not.

With this patch, all correction types implement `evaluate`.
This simplifies the implementation of the `node_evaluate`
visitor considerably and keeps the logic related
to the handling of each type of correction local
to the corrections themselves.