Store all information about the AST in FormulaAST
Before this commit, for some binary and unary operations
FormulaAST was storing opaque function pointers,
losing the information about what the operation was.
Now we always store the kind of operation as part
of the AST node.

This patch also makes treatment of different kinds
of unary and binary operations more uniform and,
during AST evaluation, removes the indirection
due to calling code through the function pointer.