Migrate to scikit-build-core (#279)
* Convert to PEP 621 (`pyproject.toml`) metadata

Move the project metadata from `setup.cfg` to `pyproject.toml`,
per PEP 621.  This is a prerequisite for switching to
`scikit-build-core`.  However, since this format is supported
by setuptools as well, convert the metadata separately to make diffs
more readable.

This also fixes including type hints (`.pyi`) in the package.

* Migrate to `scikit-build-core`

Use the modern `scikit-build-core` PEP 517 backend, rather than
the legacy `scikit-build` setuptools plugin.  `setuptools-scm` is still
used as the version provider (`scikit-build-core` supports it out of
the box).  `CMakeLists.txt` have been updated to use the standard
`SKBUILD_PROJECT_VERSION` variable to pass the version through.

* Skip building `demo` in `scikit-build-core`

Disable building the `demo` executable when invoking the PEP 517
backend, since it is not included in the resulting wheel.  Direct CMake
invocation still defaults to building it.

* Restore and test for version in C++ header

* Workaround setuptools_scm + CI issue

---------

Co-authored-by: Nick Smith <nick.smith@cern.ch>